chore: move all of the posix/windows stuff
