+++ path/_util.ts
@@ -3,7 +3,6 @@
 // Ported from https://github.com/browserify/path-browserify/
 // This module is browser compatible.
 
-import type { FormatInputPathObject } from "./_interface.ts";
 import {
   CHAR_BACKWARD_SLASH,
   CHAR_DOT,
@@ -105,57 +104,6 @@ export function normalizeString(
   return res;
 }
 
-export function _format(
-  sep: string,
-  pathObject: FormatInputPathObject,
-): string {
-  const dir: string | undefined = pathObject.dir || pathObject.root;
-  const base: string = pathObject.base ||
-    (pathObject.name || "") + (pathObject.ext || "");
-  if (!dir) return base;
-  if (base === sep) return dir;
-  if (dir === pathObject.root) return dir + base;
-  return dir + sep + base;
-}
-
-const WHITESPACE_ENCODINGS: Record<string, string> = {
-  "\u0009": "%09",
-  "\u000A": "%0A",
-  "\u000B": "%0B",
-  "\u000C": "%0C",
-  "\u000D": "%0D",
-  "\u0020": "%20",
-};
-
-export function encodeWhitespace(string: string): string {
-  return string.replaceAll(/[\s]/g, (c) => {
-    return WHITESPACE_ENCODINGS[c] ?? c;
-  });
-}
-
-export function lastPathSegment(
-  path: string,
-  isSep: (char: number) => boolean,
-  start = 0,
-): string {
-  let matchedNonSeparator = false;
-  let end = path.length;
-
-  for (let i = path.length - 1; i >= start; --i) {
-    if (isSep(path.charCodeAt(i))) {
-      if (matchedNonSeparator) {
-        start = i + 1;
-        break;
-      }
-    } else if (!matchedNonSeparator) {
-      matchedNonSeparator = true;
-      end = i + 1;
-    }
-  }
-
-  return path.slice(start, end);
-}
-
 export function stripTrailingSeparators(
   segment: string,
   isSep: (char: number) => boolean,
@@ -175,20 +123,4 @@ export function stripTrailingSeparators(
   }
 
   return segment.slice(0, end);
-}
-
-export function stripSuffix(name: string, suffix: string): string {
-  if (suffix.length >= name.length) {
-    return name;
-  }
-
-  const lenDiff = name.length - suffix.length;
-
-  for (let i = suffix.length - 1; i >= 0; --i) {
-    if (name.charCodeAt(lenDiff + i) !== suffix.charCodeAt(i)) {
-      return name;
-    }
-  }
-
-  return name.slice(0, -suffix.length);
 }
