+++ path/basename.ts
@@ -2,42 +2,7 @@
 // This module is browser compatible.
 
 import { isWindows } from "../_util/os.ts";
-import { CHAR_COLON } from "./_constants.ts";
-import {
-  assertPath,
-  isPathSeparator,
-  isPosixPathSeparator,
-  isWindowsDeviceRoot,
-  lastPathSegment,
-  stripSuffix,
-  stripTrailingSeparators,
-} from "./_util.ts";
-
-function posixBasename(path: string, suffix = ""): string {
-  const lastSegment = lastPathSegment(path, isPosixPathSeparator);
-  const strippedSegment = stripTrailingSeparators(
-    lastSegment,
-    isPosixPathSeparator,
-  );
-  return suffix ? stripSuffix(strippedSegment, suffix) : strippedSegment;
-}
-
-function windowsBasename(path: string, suffix = ""): string {
-  // Check for a drive letter prefix so as not to mistake the following
-  // path separator as an extra separator at the end of the path that can be
-  // disregarded
-  let start = 0;
-  if (path.length >= 2) {
-    const drive = path.charCodeAt(0);
-    if (isWindowsDeviceRoot(drive)) {
-      if (path.charCodeAt(1) === CHAR_COLON) start = 2;
-    }
-  }
-
-  const lastSegment = lastPathSegment(path, isPathSeparator, start);
-  const strippedSegment = stripTrailingSeparators(lastSegment, isPathSeparator);
-  return suffix ? stripSuffix(strippedSegment, suffix) : strippedSegment;
-}
+import { posixBasename, windowsBasename } from "./_basename.ts";
 
 /**
  * Return the last portion of a `path`.
@@ -47,16 +12,7 @@ function windowsBasename(path: string, suffix = ""): string {
  * @param [suffix] - suffix to remove from extracted name.
  */
 export function basename(path: string, suffix = ""): string {
-  assertPath(path);
-  if (path.length === 0) return path;
-  if (typeof suffix !== "string") {
-    throw new TypeError(
-      `Suffix must be a string. Received ${JSON.stringify(suffix)}`,
-    );
-  }
-
-  if (isWindows) {
-    return windowsBasename(path, suffix);
-  }
-  return posixBasename(path, suffix);
+  return isWindows
+    ? windowsBasename(path, suffix)
+    : posixBasename(path, suffix);
 }
