+++ path/format.ts
@@ -2,22 +2,13 @@
 // This module is browser compatible.
 
 import { isWindows } from "../_util/os.ts";
+import { posixFormat, windowsFormat } from "./_format.ts";
 import { FormatInputPathObject } from "./_interface.ts";
-import { _format } from "./_util.ts";
 
 /**
  * Generate a path from `FormatInputPathObject` object.
  * @param pathObject with path
  */
 export function format(pathObject: FormatInputPathObject): string {
-  if (pathObject === null || typeof pathObject !== "object") {
-    throw new TypeError(
-      `The "pathObject" argument must be of type Object. Received type ${typeof pathObject}`,
-    );
-  }
-
-  if (isWindows) {
-    return _format("\\", pathObject);
-  }
-  return _format("/", pathObject);
+  return isWindows ? windowsFormat(pathObject) : posixFormat(pathObject);
 }
