+++ path/parse.ts
@@ -2,265 +2,13 @@
 // This module is browser compatible.
 
 import { isWindows } from "../_util/os.ts";
-import { CHAR_COLON, CHAR_DOT } from "./_constants.ts";
-import { ParsedPath } from "./_interface.ts";
-import {
-  assertPath,
-  isPathSeparator,
-  isPosixPathSeparator,
-  isWindowsDeviceRoot,
-  stripTrailingSeparators,
-} from "./_util.ts";
-
-function posixParse(path: string): ParsedPath {
-  const ret: ParsedPath = { root: "", dir: "", base: "", ext: "", name: "" };
-  if (path.length === 0) return ret;
-  const isAbsolute = isPosixPathSeparator(path.charCodeAt(0));
-  let start: number;
-  if (isAbsolute) {
-    ret.root = "/";
-    start = 1;
-  } else {
-    start = 0;
-  }
-  let startDot = -1;
-  let startPart = 0;
-  let end = -1;
-  let matchedSlash = true;
-  let i = path.length - 1;
-
-  // Track the state of characters (if any) we see before our first dot and
-  // after any path separator we find
-  let preDotState = 0;
-
-  // Get non-dir info
-  for (; i >= start; --i) {
-    const code = path.charCodeAt(i);
-    if (isPosixPathSeparator(code)) {
-      // If we reached a path separator that was not part of a set of path
-      // separators at the end of the string, stop now
-      if (!matchedSlash) {
-        startPart = i + 1;
-        break;
-      }
-      continue;
-    }
-    if (end === -1) {
-      // We saw the first non-path separator, mark this as the end of our
-      // extension
-      matchedSlash = false;
-      end = i + 1;
-    }
-    if (code === CHAR_DOT) {
-      // If this is our first dot, mark it as the start of our extension
-      if (startDot === -1) startDot = i;
-      else if (preDotState !== 1) preDotState = 1;
-    } else if (startDot !== -1) {
-      // We saw a non-dot and non-path separator before our dot, so we should
-      // have a good chance at having a non-empty extension
-      preDotState = -1;
-    }
-  }
-
-  if (
-    startDot === -1 ||
-    end === -1 ||
-    // We saw a non-dot character immediately before the dot
-    preDotState === 0 ||
-    // The (right-most) trimmed path component is exactly '..'
-    (preDotState === 1 && startDot === end - 1 && startDot === startPart + 1)
-  ) {
-    if (end !== -1) {
-      if (startPart === 0 && isAbsolute) {
-        ret.base = ret.name = path.slice(1, end);
-      } else {
-        ret.base = ret.name = path.slice(startPart, end);
-      }
-    }
-    // Fallback to '/' in case there is no basename
-    ret.base = ret.base || "/";
-  } else {
-    if (startPart === 0 && isAbsolute) {
-      ret.name = path.slice(1, startDot);
-      ret.base = path.slice(1, end);
-    } else {
-      ret.name = path.slice(startPart, startDot);
-      ret.base = path.slice(startPart, end);
-    }
-    ret.ext = path.slice(startDot, end);
-  }
-
-  if (startPart > 0) {
-    ret.dir = stripTrailingSeparators(
-      path.slice(0, startPart - 1),
-      isPosixPathSeparator,
-    );
-  } else if (isAbsolute) ret.dir = "/";
-
-  return ret;
-}
-
-function windowsParse(path: string): ParsedPath {
-  const ret: ParsedPath = { root: "", dir: "", base: "", ext: "", name: "" };
-
-  const len = path.length;
-  if (len === 0) return ret;
-
-  let rootEnd = 0;
-  let code = path.charCodeAt(0);
-
-  // Try to match a root
-  if (len > 1) {
-    if (isPathSeparator(code)) {
-      // Possible UNC root
-
-      rootEnd = 1;
-      if (isPathSeparator(path.charCodeAt(1))) {
-        // Matched double path separator at beginning
-        let j = 2;
-        let last = j;
-        // Match 1 or more non-path separators
-        for (; j < len; ++j) {
-          if (isPathSeparator(path.charCodeAt(j))) break;
-        }
-        if (j < len && j !== last) {
-          // Matched!
-          last = j;
-          // Match 1 or more path separators
-          for (; j < len; ++j) {
-            if (!isPathSeparator(path.charCodeAt(j))) break;
-          }
-          if (j < len && j !== last) {
-            // Matched!
-            last = j;
-            // Match 1 or more non-path separators
-            for (; j < len; ++j) {
-              if (isPathSeparator(path.charCodeAt(j))) break;
-            }
-            if (j === len) {
-              // We matched a UNC root only
-
-              rootEnd = j;
-            } else if (j !== last) {
-              // We matched a UNC root with leftovers
-
-              rootEnd = j + 1;
-            }
-          }
-        }
-      }
-    } else if (isWindowsDeviceRoot(code)) {
-      // Possible device root
-
-      if (path.charCodeAt(1) === CHAR_COLON) {
-        rootEnd = 2;
-        if (len > 2) {
-          if (isPathSeparator(path.charCodeAt(2))) {
-            if (len === 3) {
-              // `path` contains just a drive root, exit early to avoid
-              // unnecessary work
-              ret.root = ret.dir = path;
-              ret.base = "\\";
-              return ret;
-            }
-            rootEnd = 3;
-          }
-        } else {
-          // `path` contains just a relative drive root, exit early to avoid
-          // unnecessary work
-          ret.root = ret.dir = path;
-          return ret;
-        }
-      }
-    }
-  } else if (isPathSeparator(code)) {
-    // `path` contains just a path separator, exit early to avoid
-    // unnecessary work
-    ret.root = ret.dir = path;
-    ret.base = "\\";
-    return ret;
-  }
-
-  if (rootEnd > 0) ret.root = path.slice(0, rootEnd);
-
-  let startDot = -1;
-  let startPart = rootEnd;
-  let end = -1;
-  let matchedSlash = true;
-  let i = path.length - 1;
-
-  // Track the state of characters (if any) we see before our first dot and
-  // after any path separator we find
-  let preDotState = 0;
-
-  // Get non-dir info
-  for (; i >= rootEnd; --i) {
-    code = path.charCodeAt(i);
-    if (isPathSeparator(code)) {
-      // If we reached a path separator that was not part of a set of path
-      // separators at the end of the string, stop now
-      if (!matchedSlash) {
-        startPart = i + 1;
-        break;
-      }
-      continue;
-    }
-    if (end === -1) {
-      // We saw the first non-path separator, mark this as the end of our
-      // extension
-      matchedSlash = false;
-      end = i + 1;
-    }
-    if (code === CHAR_DOT) {
-      // If this is our first dot, mark it as the start of our extension
-      if (startDot === -1) startDot = i;
-      else if (preDotState !== 1) preDotState = 1;
-    } else if (startDot !== -1) {
-      // We saw a non-dot and non-path separator before our dot, so we should
-      // have a good chance at having a non-empty extension
-      preDotState = -1;
-    }
-  }
-
-  if (
-    startDot === -1 ||
-    end === -1 ||
-    // We saw a non-dot character immediately before the dot
-    preDotState === 0 ||
-    // The (right-most) trimmed path component is exactly '..'
-    (preDotState === 1 && startDot === end - 1 && startDot === startPart + 1)
-  ) {
-    if (end !== -1) {
-      ret.base = ret.name = path.slice(startPart, end);
-    }
-  } else {
-    ret.name = path.slice(startPart, startDot);
-    ret.base = path.slice(startPart, end);
-    ret.ext = path.slice(startDot, end);
-  }
-
-  // Fallback to '\' in case there is no basename
-  ret.base = ret.base || "\\";
-
-  // If the directory is the root, use the entire root as the `dir` including
-  // the trailing slash if any (`C:\abc` -> `C:\`). Otherwise, strip out the
-  // trailing slash (`C:\abc\def` -> `C:\abc`).
-  if (startPart > 0 && startPart !== rootEnd) {
-    ret.dir = path.slice(0, startPart - 1);
-  } else ret.dir = ret.root;
-
-  return ret;
-}
+import type { ParsedPath } from "./_interface.ts";
+import { posixParse, windowsParse } from "./_parse.ts";
 
 /**
  * Return a `ParsedPath` object of the `path`.
  * @param path to process
  */
 export function parse(path: string): ParsedPath {
-  assertPath(path);
-
-  if (isWindows) {
-    return windowsParse(path);
-  }
-  return posixParse(path);
+  return isWindows ? windowsParse(path) : posixParse(path);
 }
